Move useAsyncLoader to /utilities/hooks with the other hooks
